Separate ArtifactReporter to subt/artf_reporter.py
